Fix colors button and move the axios.js into states folder
